Move router root from entry to new component.
